backend/requirements.txt
asgiref==3.5.0
boto3==1.24.94
botocore==1.27.94
certifi==2022.9.24
distlib==0.3.6
Django==3.2.12
django-cors-headers==3.11.0
django-environ==0.8.1
django-storages==1.13.1
djangorestframework==3.13.1
filelock==3.8.0
jmespath==1.0.1
npm==0.1.1
optional-django==0.1.0
Pillow==9.2.0
pipenv==2022.10.12
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
whitenoise==6.2.0
